feat: 모의투자 매수 기능 구현

+++ back/src/main/java/com/stock/survive/dto/MyPageDto.java
@@ -17,7 +17,7 @@ public class MyPageDto {
     private String nickname;
     private boolean cancel;
     private Integer totalReward;
-    private Integer cash;
+    private Long cash;
     private String createdAt;
     private String avatarUrl;
